add get_user_from_email to register_user
