+++ backend/ops_api/tests/auth/test_oidc_controller.py
@@ -43,8 +43,6 @@ def test_auth_post_succeeds_creates_event(client, loaded_db, mocker):
     mock_session = mocker.MagicMock()
     mock_cm.return_value.__enter__.return_value = mock_session
 
-    m1 = mocker.patch("ops_api.ops.auth.service._get_token_and_user_data_from_oauth_provider")
-    m1.return_value = ({"access_token": "admin_user"}, {})
     m2 = mocker.patch("ops_api.ops.auth.service._get_token_and_user_data_from_internal_auth")
     user = mocker.MagicMock()
     user.to_dict.return_value = {}
@@ -62,8 +60,6 @@ def test_auth_post_succeeds_creates_event(client, loaded_db, mocker):
 
 def test_login_succeeds_with_active_status(client, loaded_db, mocker):
     # setup mocks
-    m1 = mocker.patch("ops_api.ops.auth.service._get_token_and_user_data_from_oauth_provider")
-    m1.return_value = ({"access_token": "admin_user"}, {})
     m2 = mocker.patch("ops_api.ops.auth.service._get_token_and_user_data_from_internal_auth")
     user = mocker.MagicMock()
     user.to_dict.return_value = {}
